Add CircuitSiEPICFormatter
