+++ simphony/plugins/siepic/tests/test_siepic.py
@@ -6,7 +6,7 @@ import os
 
 import pytest
 
-from simphony.plugins.siepic.parser import load_spi
+from simphony.plugins.siepic import load_spi_from_file
 
 # ==============================================================================
 # Test the parser
@@ -548,19 +548,19 @@ def test_EBeam_sequoiap_A_v2():
         "EBeam_sequoiap_A_v2",
         "EBeam_sequoiap_A_v2_main.spi",
     )
-    res = load_spi(filename)
+    res = load_spi_from_file(filename)
     assert res == EBeam_sequoiap_A_v2_result
 
 
 def test_MZI4():
     filename = os.path.join(os.path.dirname(__file__), "spice", "MZI4", "MZI4_main.spi")
-    res = load_spi(filename)
+    res = load_spi_from_file(filename)
     assert res == MZI4_result
 
 
 def test_top():
     filename = os.path.join(os.path.dirname(__file__), "spice", "top", "top_main.spi")
-    res = load_spi(filename)
+    res = load_spi_from_file(filename)
     assert res == top_result
 
 
@@ -571,6 +571,6 @@ def test_top():
 # filename = os.path.join('tests', 'spice', 'MZI4', 'MZI4_main.spi')
 # filename = os.path.join('tests', 'spice', 'EBeam_sequoiap_A_v2', 'EBeam_sequoiap_A_v2_main.spi')
 # filename = os.path.join('tests', 'spice', 'top', 'top_main.spi')
-# data = load_spi(filename)
+# data = load_spi_from_file(filename)
 # from simphony.plugins.siepic.builders import build_circuit
 # build_circuit(data, 'simphony.libraries.siepic')
